Change potential bad casts to be saturated and remove redundant casts (#2433)

+++ src/main/java/com/gregtechceu/gtceu/api/capability/recipe/FluidRecipeCapability.java
@@ -288,7 +288,7 @@ public class FluidRecipeCapability extends RecipeCapability<SizedFluidIngredient
                 }
             }
             if (available >= needed) {
-                int ratio = (int) Math.min(parallelAmount, available / needed);
+                int ratio = (int) Math.min(parallelAmount, (float) available / needed);
                 if (ratio < minMultiplier) {
                     minMultiplier = ratio;
                 }
